refactor(recommendations): move product matching out of rag
