+++ src/lib/agent/tools/select-products.ts
@@ -22,7 +22,7 @@ import {
   type CareBalanceToolRow,
 } from "@/lib/agent/tools/care-balance-context"
 import { applyProductMemoryConstraints } from "@/lib/chat-runtime/user-memory"
-import type { MatchedProduct } from "@/lib/rag/product-matcher"
+import type { MatchedProduct } from "@/lib/product-matching/matcher"
 import type { UserMemoryContext } from "@/lib/chat-runtime/user-memory"
 import type {
   BondbuilderRecommendationMetadata,
