+++ src/lib/rag/conditioner-decision.ts
@@ -14,7 +14,7 @@ import type {
   ConditionerRecommendationMetadata,
   HairProfile,
 } from "@/lib/types"
-import type { MatchedProduct } from "@/lib/rag/product-matcher"
+import type { MatchedProduct } from "@/lib/product-matching/matcher"
 
 const CONDITIONER_CLARIFICATION_QUESTIONS: Record<ConditionerProfileField, string> = {
   thickness: "Ist dein Haar eher fein, mittel oder dick?",
